+++ just_audio_background/lib/just_audio_background.dart
@@ -523,6 +523,8 @@ class _PlayerAudioHandler extends BaseAudioHandler
       ConcatenatingInsertAllRequest request) async {
     final cat = _source!.findCat(request.id)!;
     cat.children.insertAll(request.index, request.children);
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -533,6 +535,8 @@ class _PlayerAudioHandler extends BaseAudioHandler
       ConcatenatingRemoveRangeRequest request) async {
     final cat = _source!.findCat(request.id)!;
     cat.children.removeRange(request.startIndex, request.endIndex);
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -544,6 +548,8 @@ class _PlayerAudioHandler extends BaseAudioHandler
     final cat = _source!.findCat(request.id)!;
     cat.children
         .insert(request.newIndex, cat.children.removeAt(request.currentIndex));
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();